Migrating notifications to component
